requirements.txt
torch
torchvision
pillow
numpy
typing
scikit-image
transformers>=4.39.1
kornia
fastapi
uvicorn
python-multipart
